add option to edit a member, fix the option to put a number bigger than int , and show a every loan ever made and  alert if someone has loaned this book before

+++ libaryApp/DataStructs.cs
@@ -102,7 +102,7 @@ namespace libaryApp
     {
         public string memberName { get; set; }
         public int MemberID { get; set; }
-        public int PersonID { get; set; }
+        public long PersonID { get; set; }
         public string Phone { get; set; }
         public string Adress { get; set; }
